Added functionality for exporting and importing chromosomes
The popup.html had to be changed a bit because the maximum height of
popups is 600 px. If the Developer header is left it needs more than 600
px height.

toJSON() is called when JSON.stringify() is called which leads to an
endless loop. Either rename the function or just leave it removed as
it's not used anywhere.