Betting limit, no betting on exhibitions, load chat instead of whole site
Fixes #64 and #66

You can now toggle if betting on exhibitions is enabled.

With the checkbox and input field you can input limits when the bot
should not bet anymore.

Instead of the whole twitch site only the chat is loaded.

Made the chat script more reliable by changing from onload to
onreadystate.change.

Fixed the month on the record and chromosome files, was 1 month too low.
(January = 0)