Added dates to records, reload chat every hour, changed match code format
A date was added to each record so you know how old records are.
If the bot decides to bet nothing 1 dollar is bet instead. (Useful for
NEW tier)
The twitch tab reloads every hour because when it times out it will
never reload.
The _ in the match code was replaced with | to make it more
understandable.