winPercentageWeight is split among the characters
If a character has a win percentage of 60 % and the other of 30 % then
one character will get 2/3 of the weight and the other one 1/3. This is
important for S tier fights where the win rates are pretty equal.

The earned salt slighty increases with this change. Betting levels can
be raised in the future because the average betting amount decreased
with this change.