Read mode from the footer, fixed a bug with resizing the window with video disabled
Saltbot now doesn't rely on the chat to get the mode but reads the
information from the footer when betting is available.

When resizing the window while the stream was disabled with toggle video
it would mess up the div layout. Emptying the div instead of removing it
preserves the page layout when resizing.