Config page, added option to all in on touneys
Resolves #87

The cogwheel in the popup window leads to a new page with the options:

* Exhibition betting
* Going all in at tournaments
* Setting a tournament-specific limit
* Listing the player rankings

The old betting limit is now only used for non-tournament matches.

The messages for any changed settings still appear in the log of the
saltybet tab.

The whole application code has been reformatted so indentations are
correct.