Fixed exhibition tiers and commas in character names
Fixes #88.

For exhibitions the tier is set to U again if no tier was found in the
waifu message.

Commas are now also replaced in the waifu4u parser.
Made config.html HTML5-conform.

Swapped the log messages of team match and exhibition deactivated so the
messages can be grouped to 25 exhibition deactivated log messages in the
console log.