Tinkering chormosome default weights and handling.
The original weights might force the sim solutions into a particular
locked direction, there might be a reason for it but i have no
documentation to prove it works.  The goal of this defaults change to
start as blank as possible and let nature find its way.
Testing reverting normalization to 0..1 values,   i am unsure how the
winning heuristics process the scale of the weights, but if the these
(weights) versus (100xweights) changes the final bet and its confidence,
something is unclear in the strategy class i need to study.  Recommend
editing the printout function to x100 the weights.