improvements by mutation scale tweaking
Discovered a small  'mutationScale' value lead to finer refinements of
correct % and total earnings.  This benefits the sensitivity of the
chromosome weights.  This change may also benefit the sim as new match
data is added at runtime. Also observed that there are weight sets that
together  affect win % or winnings improvement (observing generation
change), my guess some weights correlated better for odds, and others
correlated for confidence and winning, but to find it a data
classification program will need to analyze.
Problem with change, sim generations count required is now in the 50-150
range.