Develop (#13)
* Fixed restarting of crashed saltybet tabs

Because of the asynchronous way of querying tabs in Chrome the variables had other values then expected. Fixed it by closing all saltybet tabs and using a global variable with time-delayed setting so tabs aren't restarted multiple times in a row.

Related to #113

* Renamed function setRestarted used in keepAlive mode

* Gobal cleanup and error checks (#10)

* Added, prediction console log for character and conf.

* modification. tourney & scientist mode, no longer bets bailout without the option.

Tourney bets are now conservative when scientist wants to be.

* fix error printing

*  balance/bailout behavior fix.

* Changed. tourney betting scale for sci mode & not all-in mode (#11)

* implement better  scale range for tourney betting (#12)

* (tourney only) Default confidence range (0.5, 1) is optionally remapped to a new range.
This should help solve scientist mode bets ending tourneys in bailout.