Bug fix new pool start (#15)
* fix starting pool breeding selection

fix, breeding would use the starting class chromosomes, which are not
normalized, and getting an extremely wide range of starting values.

* changes managing breeding differences: pop size, mutation scale

scale between best and worst weights in the populations was cause the
appearance of dominate weights as quick as gen one. Mutations scaled
higher hopefully to counter this effect too.

* disabled best/worst breed rule

purpose to be decided.